Added permission checks to invite commands, fixed invite bugs, and allowed inviter to cancel invites

+++ src/main/kotlin/dev/rayzr/gameboi/command/InviteCommands.kt
@@ -7,11 +7,30 @@ import dev.rayzr.gameboi.game.fight.FightGame
 import dev.rayzr.gameboi.game.hangman.HangmanGame
 import dev.rayzr.gameboi.game.twenty48.Twenty48Game
 import dev.rayzr.gameboi.manager.InviteManager
+import net.dv8tion.jda.api.Permission
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent
 import java.util.concurrent.TimeUnit
 
+fun checkPermissions(event: GuildMessageReceivedEvent): Boolean {
+    if (!event.guild.selfMember.hasPermission(event.channel, Permission.MESSAGE_WRITE)) {
+        // Can't do anything about this, oh well
+        return false
+    }
+
+    if (!event.guild.selfMember.hasPermission(event.channel, Permission.MESSAGE_MANAGE, Permission.MESSAGE_ATTACH_FILES, Permission.MESSAGE_EMBED_LINKS)) {
+        event.channel.sendMessage(":x: This bot is missing the permissions required to manage/delete messages, attach images, and embed links!").queue()
+        return false
+    }
+
+    return true
+}
+
 object Connect4Invite : Command("connect4", "Invites a player to play Connect4 with you!", "connect4 <other>") {
     override fun handle(event: GuildMessageReceivedEvent, args: List<String>) {
+        if (!checkPermissions(event)) {
+            return
+        }
+
         if (event.message.mentionedMembers.size < 1) {
             event.channel.sendMessage(":x: Please mention the user you would like to play with!").queue {
                 it.textChannel.deleteMessages(listOf(it, event.message)).queueAfter(Gameboi.errorLife, TimeUnit.MILLISECONDS)
@@ -27,6 +46,10 @@ object Connect4Invite : Command("connect4", "Invites a player to play Connect4 w
 
 object FightInvite : Command("fight", "Invites a player to play Fight with you!", "fight <other>") {
     override fun handle(event: GuildMessageReceivedEvent, args: List<String>) {
+        if (!checkPermissions(event)) {
+            return
+        }
+
         if (event.message.mentionedMembers.size < 1) {
             event.channel.sendMessage(":x: Please mention the user you would like to play with!").queue {
                 it.textChannel.deleteMessages(listOf(it, event.message)).queueAfter(Gameboi.errorLife, TimeUnit.MILLISECONDS)
@@ -42,12 +65,20 @@ object FightInvite : Command("fight", "Invites a player to play Fight with you!"
 
 object Twenty48Invite: Command("2048", "Starts a 2048 game.", "2048") {
     override fun handle(event: GuildMessageReceivedEvent, args: List<String>) {
+        if (!checkPermissions(event)) {
+            return
+        }
+
         InviteManager.singlePlayer(event.message, Player[event.author], Twenty48Game)
     }
 }
 
 object HangmanInvite: Command("hangman", "Starts a hangman game.", "hangman") {
     override fun handle(event: GuildMessageReceivedEvent, args: List<String>) {
+        if (!checkPermissions(event)) {
+            return
+        }
+
         InviteManager.singlePlayer(event.message, Player[event.author], HangmanGame)
     }
 }
